Change the calculation of max for service metrics
The current calculation does not work and always returns 0. As the mean works correctly, I assume the max function on Timer does not return a useful value. As the code iterates through all requests anyway, it can also use the total time of one request and then take the largest value.